requirements.txt
django==5.2
python-decouple==3.8
psycopg==3.2.2
gunicorn==23.0.0
dj-database-url==2.2.0
whitenoise==6.7.0
